Move mathematical functions to MathUtils
